+++ test/dummy/config/routes.rb
@@ -1,3 +1,5 @@
 Dummy::Application.routes.draw do
   resources :users
+
+  root to: "bootstrap#form"
 end